pyproject.toml
[project]
name = "pydefi-simulator"
version = "0.1.0"
description = "DeFi simulator for testing lending and borrowing strategies."
readme = "README.md"
keywords = ["cryptocurrency", "defi", "simulator"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
authors = [
    { email = "robin.nogues@google.com" },
    { name = "Robin Nogues" },
]
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"

[dependency-groups]
dev = [
    "flake8>=7.1.1",
    "mypy>=1.14.1",
    "pip>=25.0",
    "ruff>=0.9.4",
    "types-pyyaml>=6.0.12.20241230",
    "types-requests>=2.32.0.20241016",
    "uv>=0.5.26",
]
